Moved reverse nested code into the nested package Removed automatic path resolving If no path is defined then the reverse_nested agg always joins back to the root / main document level.
