+++ src/main/java/org/elasticsearch/search/aggregations/bucket/nested/ReverseNested.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.search.aggregations.bucket.reversenested;
+package org.elasticsearch.search.aggregations.bucket.nested;
 
 import org.elasticsearch.search.aggregations.bucket.SingleBucketAggregation;
